sweepの指定を2000x5000みたいな感じで指定したい
