+++ tools/Ptolemy.Libra/Libra.cs
@@ -4,33 +4,43 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using Ptolemy.Libra.Request;
+using Ptolemy.Repository;
 
 namespace Ptolemy.Libra {
     public class Libra {
         private readonly CancellationToken token;
-        private readonly LibraRequest request;
 
         /// <summary>
         /// 
         /// </summary>
         /// <param name="token"></param>
-        /// <param name="request"></param>
-        public Libra(CancellationToken token, LibraRequest request) {
+        public Libra(CancellationToken token) {
             this.token = token;
-            this.request = request;
         }
 
         /// <summary>
         /// requestに従って数え上げる
         /// </summary>
         /// <returns>Expression,結果のペアリスト</returns>
-        public Tuple<string,long>[] Run() {
+        public Tuple<string,long>[] Run(LibraRequest request) {
             try {
                 var delegates = request.BuildFilter();
                 var signals = request.SignalList;
 
                 if (!signals.Any()) throw new LibraException("Conditions have no signals");
 
+                var db = new ReadOnlyRepository(request.SqliteFile);
+
+                var result = request.IsSplitWithSeed switch {
+                    true => db.Aggregate(token, signals, delegates,
+                        Range(request.SeedStart, request.SeedEnd).ToList(),
+                        request.SweepStart, request.SweepEnd, LibraRequest.GetKey),
+                    //false => db.Aggregate(token, signals, delegates,request.SeedStart, )
+                    };
+
+                if (request.IsSplitWithSeed) {
+                    
+                }
 
                 return null;
             }
@@ -40,6 +50,10 @@ namespace Ptolemy.Libra {
             catch (Exception e) {
                 throw new LibraException($"Unknown error has occured\n\t-->{e}");
             }
+        }
+
+        private static IEnumerable<long> Range(long start, long end) {
+            for (var l = start; l <= end; l++) yield return l;
         }
     }
